+++ backend/services/budget.service.go
@@ -2,10 +2,12 @@ package services
 
 import (
 	"database/sql"
+	"time"
 
 	"github.com/google/uuid"
 	"github.com/rahulcodepython/finance-tracker-backend/backend/models"
 	"github.com/rahulcodepython/finance-tracker-backend/backend/repository"
+	"github.com/rahulcodepython/finance-tracker-backend/backend/utils"
 )
 
 func CreateBudget(userID uuid.UUID, name string, amount float64, db *sql.DB) (*models.Budget, error) {
@@ -34,8 +36,13 @@ func UpdateBudget(id uuid.UUID, name string, amount float64, db *sql.DB) (*model
 		return nil, err
 	}
 
+	if budget == nil {
+		return nil, sql.ErrNoRows
+	}
+
 	budget.Name = name
 	budget.Amount = amount
+	budget.UpdatedAt = time.Now().In(utils.LOC)
 
 	err = repository.UpdateBudget(budget, db)
 	if err != nil {
@@ -46,15 +53,21 @@ func UpdateBudget(id uuid.UUID, name string, amount float64, db *sql.DB) (*model
 }
 
 func DeleteBudget(id uuid.UUID, db *sql.DB) error {
+	budget, err := repository.GetBudgetByID(id, db)
+	if err != nil {
+		return err
+	}
+
+	if budget == nil {
+		return sql.ErrNoRows
+	}
+
 	return repository.DeleteBudget(id, db)
 }
 
 func CheckBudgetExistsById(id uuid.UUID, db *sql.DB) (bool, error) {
 	budget, err := repository.GetBudgetByID(id, db)
 	if err != nil {
-		if err == sql.ErrNoRows {
-			return false, nil
-		}
 		return false, err
 	}
