+++ backend/services/recurring.transaction.service.go
@@ -10,15 +10,50 @@ import (
 	"github.com/rahulcodepython/finance-tracker-backend/backend/utils"
 )
 
-func CreateRecurringTransaction(userID uuid.UUID, accountID uuid.UUID, categoryID uuid.UUID, description string, amount float64, transactionType models.TransactionType, recurringFrequency models.RecurringFrequency, recurringDate int, db *sql.DB) (*models.RecurringTransaction, error) {
+func CreateRecurringTransaction(userID uuid.UUID, accountID uuid.UUID, categoryID uuid.UUID, budgetID uuid.NullUUID, description string, amount float64, note sql.NullString, recurringFrequency models.RecurringFrequency, recurringDate int, db *sql.DB) (*models.RecurringTransaction, error) {
+	category, err := repository.GetCategoryByID(categoryID, db)
+	if err != nil {
+		return nil, err
+	}
+
+	if category == nil {
+		return nil, sql.ErrNoRows
+	}
+
+	transactionType := models.TransactionType(category.Type)
+
+	// Update account balance
+	account, err := repository.GetAccountByID(accountID, db)
+	if err != nil {
+		return nil, err
+	}
+
+	if account == nil {
+		return nil, sql.ErrNoRows
+	}
+
+	// Update budget if provided
+	if budgetID.Valid {
+		budget, err := repository.GetBudgetByID(budgetID.UUID, db)
+		if err != nil {
+			return nil, err
+		}
+
+		if budget == nil {
+			return nil, sql.ErrNoRows
+		}
+	}
+
 	recurringTransaction := &models.RecurringTransaction{
 		ID:                 uuid.New(),
 		UserID:             userID,
 		AccountID:          accountID,
-		CategoryID:         uuid.NullUUID{UUID: categoryID, Valid: true},
+		CategoryID:         categoryID,
+		BudgetID:           budgetID,
 		Description:        description,
 		Amount:             amount,
 		Type:               transactionType,
+		Note:               note,
 		RecurringFrequency: recurringFrequency,
 		RecurringDate:      recurringDate,
 		CreatedAt:          time.Now().In(utils.LOC),
@@ -36,17 +71,54 @@ func GetRecurringTransactions(userID uuid.UUID, db *sql.DB) ([]models.RecurringT
 	return repository.GetRecurringTransactionsByUserID(userID, db)
 }
 
-func UpdateRecurringTransaction(id uuid.UUID, accountID uuid.UUID, categoryID uuid.UUID, description string, amount float64, transactionType models.TransactionType, recurringFrequency models.RecurringFrequency, recurringDate int, db *sql.DB) (*models.RecurringTransaction, error) {
+func UpdateRecurringTransaction(id uuid.UUID, accountID uuid.UUID, categoryID uuid.UUID, budgetID uuid.NullUUID, description string, amount float64, note sql.NullString, recurringFrequency models.RecurringFrequency, recurringDate int, db *sql.DB) (*models.RecurringTransaction, error) {
 	recurringTransaction, err := repository.GetRecurringTransactionByID(id, db)
 	if err != nil {
 		return nil, err
 	}
 
+	if recurringTransaction == nil {
+		return nil, sql.ErrNoRows
+	}
+
+	category, err := repository.GetCategoryByID(categoryID, db)
+	if err != nil {
+		return nil, err
+	}
+
+	if category == nil {
+		return nil, sql.ErrNoRows
+	}
+
+	transactionType := models.TransactionType(category.Type)
+
+	account, err := repository.GetAccountByID(accountID, db)
+	if err != nil {
+		return nil, err
+	}
+
+	if account == nil {
+		return nil, sql.ErrNoRows
+	}
+
+	if budgetID.Valid {
+		budget, err := repository.GetBudgetByID(budgetID.UUID, db)
+		if err != nil {
+			return nil, err
+		}
+
+		if budget == nil {
+			return nil, sql.ErrNoRows
+		}
+	}
+
 	recurringTransaction.AccountID = accountID
-	recurringTransaction.CategoryID = uuid.NullUUID{UUID: categoryID, Valid: true}
+	recurringTransaction.CategoryID = categoryID
+	recurringTransaction.BudgetID = budgetID
 	recurringTransaction.Description = description
 	recurringTransaction.Amount = amount
 	recurringTransaction.Type = transactionType
+	recurringTransaction.Note = note
 	recurringTransaction.RecurringFrequency = recurringFrequency
 	recurringTransaction.RecurringDate = recurringDate
 	recurringTransaction.UpdatedAt = time.Now().In(utils.LOC)
@@ -59,5 +131,14 @@ func UpdateRecurringTransaction(id uuid.UUID, accountID uuid.UUID, categoryID uu
 }
 
 func DeleteRecurringTransaction(id uuid.UUID, db *sql.DB) error {
+	recurringTransaction, err := repository.GetRecurringTransactionByID(id, db)
+	if err != nil {
+		return err
+	}
+
+	if recurringTransaction == nil {
+		return sql.ErrNoRows
+	}
+
 	return repository.DeleteRecurringTransaction(id, db)
 }
